Workaround a windows test failure.
In Python 3.8, os.path.realpath() was changed to resolve symbolic
links and junctions on Windows. In my setup, the pstore source tree
lives on a network drive. This was resolved to an UNC path which
caused later code to fail.

The hacky fix (in my case at least) is to use os.path.abspath() instead.