Use the small_vector<> class from peejay. (#22)
The pstore small_vector<> class was only ever a very basic implementation. It didn't support types without a default constructor and wanted to default-construct instances in the "small" container whether they were needed or not.

The peejay implementation is substantially more complete, so it makes sense to use it instead.